+++ apps/api/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   `postResolveObjects` returns a permanent redirect to `getResolveObjects`.
+-   `getResolveObjects` does not include `POST` as an allowed method.
+
 ### Deprecated
 
 ### Fixed