+++ apps/api/CHANGELOG.md
@@ -24,6 +24,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.6.0] - 2023-05-18
+
+### Changed
+
+-   `getNodes` now returns `308` responses for UUIDs that have been sunk as synonyms of other nodes.
+
+## [2.5.1] - 2023-05-14
+
+### Changed
+
+-   Upgrades:
+    -   Node.js 18.
+    -   TypeScript 5.0.4.
+
+## [2.5.0] - 2023-05-03
+
+### Changed
+
+-   `postResolveObjects` returns a permanent redirect to `getResolveObjects`.
+-   `getResolveObjects` does not include `POST` as an allowed method.
+
+### Fixed
+
+-   Validations were always of type `BAD_REQUEST_BODY` even when the field was in the parameters.
+
+### Security
+
 ## [2.4.1] - 2023-04-30
 
 ### Fixed